refactor(hook): add useModelSettingsUrl hook
