+++ x-pack/platform/plugins/shared/streams_app/public/components/sig_events/significant_events_discovery/components/settings/tab.tsx
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useCallback, useState } from 'react';
 import {
   EuiBadge,
   EuiBottomBar,
@@ -25,7 +25,6 @@ import {
   EuiTextArea,
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { MANAGEMENT_APP_LOCATOR } from '@kbn/deeplinks-management/constants';
 import { OBSERVABILITY_STREAMS_SIG_EVENTS_INDEX_PATTERNS } from '@kbn/management-settings-ids';
 import { DEFAULT_INDEX_PATTERNS } from '@kbn/streams-schema';
 import {
@@ -33,26 +32,13 @@ import {
   MIN_EXTRACTION_INTERVAL_HOURS,
 } from '@kbn/streams-plugin/common';
 import { useKibana } from '../../../../../hooks/use_kibana';
+import { useModelSettingsUrl } from '../../../../../hooks/use_model_settings_url';
 import { getFormattedError } from '../../../../../util/errors';
 import { useContinuousExtractionSettings } from './use_continuous_extraction_settings';
 
 export function SettingsTab() {
-  const {
-    dependencies: {
-      start: { share },
-    },
-    core,
-  } = useKibana();
-
-  const modelSettingsUrl = useMemo(() => {
-    const managementLocator = share.url.locators.get(MANAGEMENT_APP_LOCATOR);
-    return (
-      managementLocator?.getRedirectUrl({
-        sectionId: 'modelManagement',
-        appId: 'model_settings',
-      }) ?? ''
-    );
-  }, [share.url.locators]);
+  const { core } = useKibana();
+  const modelSettingsUrl = useModelSettingsUrl();
 
   const [savedIndexPatterns, setSavedIndexPatterns] = useState<string>(() =>
     core.settings.client.get<string>(
