+++ x-pack/platform/plugins/shared/streams_app/public/components/sig_events/stream_detail_significant_events_view/empty_state/index.tsx
@@ -17,10 +17,10 @@ import {
   EuiText,
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { MANAGEMENT_APP_LOCATOR } from '@kbn/deeplinks-management/constants';
 import { GEN_AI_SETTINGS_DEFAULT_AI_CONNECTOR } from '@kbn/management-settings-ids';
-import React, { useMemo } from 'react';
+import React from 'react';
 import { useKibana } from '../../../../hooks/use_kibana';
+import { useModelSettingsUrl } from '../../../../hooks/use_model_settings_url';
 import noSigEventsImage from './no_sig_events.svg';
 
 const NO_DEFAULT_CONNECTOR = 'NO_DEFAULT_CONNECTOR';
@@ -38,25 +38,12 @@ export function EmptyState({
   onCancelGenerationClick: () => void;
   onGenerateSuggestionsClick: () => void;
 }) {
-  const {
-    dependencies: {
-      start: { share },
-    },
-    core,
-  } = useKibana();
+  const { core } = useKibana();
   const defaultConnector = core.uiSettings.get<string>(GEN_AI_SETTINGS_DEFAULT_AI_CONNECTOR);
+  const modelSettingsUrl = useModelSettingsUrl();
 
   const isDefaultAiConnectorMissing =
     !defaultConnector || defaultConnector === NO_DEFAULT_CONNECTOR;
-  const genAiSettingsUrl = useMemo(() => {
-    const managementLocator = share.url.locators.get(MANAGEMENT_APP_LOCATOR);
-    return (
-      managementLocator?.getRedirectUrl({
-        sectionId: 'modelManagement',
-        appId: 'model_settings',
-      }) ?? ''
-    );
-  }, [share.url.locators]);
 
   return (
     <EuiEmptyPrompt
@@ -96,7 +83,7 @@ export function EmptyState({
               >
                 <p>
                   {NO_DEFAULT_CONNECTOR_CALLOUT_DESCRIPTION}{' '}
-                  <EuiLink href={genAiSettingsUrl} external>
+                  <EuiLink href={modelSettingsUrl} external>
                     {NO_DEFAULT_CONNECTOR_CALLOUT_LINK_LABEL}
                   </EuiLink>
                 </p>
